Remove usages of is_literal_type
The type trait has been deprecated in C++17 and removed in C++20
because it offers no help in generic programming.

Its usage is not needed in strong_typedef, where it was used to SFINAE
out specializations that were otherwise identical to the primary
template, except that they declared constexpr functions instead of
non-constexpr. But this is unnecessary: a function template can always
be marked as constexpr, the compiler will simply "ignore" it [1] if a
given instantation cannot actually be constexpr.

Hence just leave the constexpr code, drop the specializations and the
related SFINAE.

[1] https://eel.is/c++draft/dcl.constexpr#7.sentence-1

Fixes #8